Move ResolveOAuthBearerHandler to SaslOauthbearerConfigHelper
